+++ src/main/java/com/fpsmod/ottersciv/reward/RewardReason.java
@@ -1,0 +1,6 @@
+package com.fpsmod.ottersciv.reward;
+
+public enum RewardReason {
+    BLOCK_BREAK,
+    MOB_KILL
+}